refactor: move duty worker and related
