+++ crates/sequencer/src/duty/errors.rs
