+++ crates/sequencer/src/duty/extractor.rs
@@ -9,7 +9,10 @@ use strata_state::{
 use strata_storage::L2BlockManager;
 
 use super::types::{BlockSigningDuty, Duty};
-use crate::{checkpoint::CheckpointHandle, errors::Error, types::BatchCheckpointDuty};
+use crate::{
+    checkpoint::CheckpointHandle,
+    duty::{errors::Error, types::BatchCheckpointDuty},
+};
 
 /// Extracts new duties given a consensus state and a identity.
 pub fn extract_duties(
